feat: Split representation
