+++ PunkAPI/Classes/Content/Beer.swift
@@ -35,39 +35,3 @@ public struct Beer: Codable {
     var brewersTips: String?
     var contributedBy: String?
 }
-
-struct Volume: Codable {
-    var value: Float?
-    var unit: String?
-}
-
-struct Recipe: Codable {
-    
-    struct Ingredient: Codable {
-        var name: String?
-        var amount: Quantity?
-    }
-    
-    var malt: [Ingredient]
-    var hops: [Ingredient]
-    var yeast: String?
-}
-
-struct Method: Codable {
-
-    
-
-    struct Step: Codable {
-        var temp: Quantity?
-        var duration: Float?
-    }
-
-    var mashTemp: [Step]?
-    var fermentation: Quantity?
-    var twist: String?
-}
-
-struct Quantity: Codable {
-    var value: Float?
-    var unit: String?
-}
